STAGE 1 of refactoring StorageHandler: file copying

+++ catroid/src/main/java/org/catrobat/catroid/ui/dialogs/NewSpriteDialog.java
@@ -223,7 +223,7 @@ public class NewSpriteDialog extends DialogFragment {
 						newObjectName = new File(lookUri.toString()).getName();
 						break;
 					case REQUEST_TAKE_PICTURE:
-						lookUri = UtilCamera.rotatePictureIfNecessary(lookUri, getString(R.string.default_look_name));
+						lookUri = UtilCamera.rotatePictureIfNecessary(lookUri);
 						break;
 					case REQUEST_MEDIA_LIBRARY:
 						lookUri = Uri.parse(data.getStringExtra(WebViewActivity.MEDIA_FILE_PATH));
@@ -404,8 +404,14 @@ public class NewSpriteDialog extends DialogFragment {
 				newLookFile = new File(lookUri.getPath());
 			} else {
 				lookData = new LookData();
-				newLookFile = StorageHandler.getInstance().copyImage(projectManager.getCurrentProject().getName(),
-						projectManager.getCurrentScene().getName(), lookUri.getPath(), null);
+				newLookFile = StorageHandler.copyFile(lookUri.getPath(),
+						projectManager.getCurrentScene().getSceneImageDirectoryPath(),
+						projectManager.getFileChecksumContainer());
+
+				if(newLookFile == null) {
+					return false;
+				}
+
 				if (lookUri.getPath().contains(Constants.TMP_LOOKS_PATH)) {
 					File oldFile = new File(lookUri.getPath());
 					oldFile.delete();
